chore: Re-organize e2e tests in anticipation to smoke tests
