+++ .gitignore
@@ -11,3 +11,4 @@ tests/*.log
 tests/__pycache__
 
 e2e/screenshots
+e2e/tests/__pycache__
