+++ e2e/run_tests.py
@@ -1,0 +1,8 @@
+import unittest
+
+if __name__ == '__main__':
+    loader = unittest.TestLoader()
+    suite = loader.discover('tests')
+
+    runner = unittest.TextTestRunner(verbosity=2)
+    runner.run(suite)
